Renaming maintenance_events to maint_notifications part 2: renaming the files.
